+++ src/app/job-listings/job-listings.component.ts
@@ -1,12 +1,21 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import { Listing } from './../listing';
+import { LISTINGS } from './../mock-listings';
 
 @Component({
   selector: 'surely-job-listings',
   templateUrl: './job-listings.component.html',
   styleUrls: ['./job-listings.component.sass'],
 })
-export class JobListingsComponent {
-  messages: number = 3;
-  notifications: number = 5;
-  email: string = 'indeed@example.org';
+export class JobListingsComponent implements OnInit {
+  listings = LISTINGS;
+  selectedListing?: Listing;
+
+  constructor() {}
+
+  ngOnInit() {}
+
+  onSelect(listing: Listing): void {
+    this.selectedListing = listing;
+  }
 }